Please Replace code of website.html with this version
This version introduces the concept of linear gradient backgrounds, which is rather popular in today web standards and also simple to learn. It also makes the navbar hover more obvious with contrasting colors. The 'Source:' part in each section is meant to show how hyperlinks work in html with the <a> tag. The Search button has been replaced with search icon to show html unicode integrations with icons. The search box has now been added with the placeholder called 'Search Box' as to make it look more clean with the simple 'placeholder=""' attribute. Hope this version is recognised too alongside the previous version. I tried to make a modern website with the same theme to make it simple with only basic HTML/CSS.